+++ inc/pages/RunPage.php
@@ -53,8 +53,7 @@ class RunPage extends Page {
 			. '<div class="well">'
 				. '<h3>History</h3>'
 				. '<ul id="history"></ul>'
-			. '</div>'
-			. '<div id="iframes"></div>';
+			. '</div>';
 
 		return $html;
 	}